ui-tests issue#3472
